+++ packages/typespec-azure-resource-manager/CHANGELOG.md
@@ -1,6 +1,26 @@
 # Change Log - @azure-tools/typespec-azure-resource-manager
 
-This log was last generated on Wed, 06 Dec 2023 19:47:28 GMT and should not be manually modified.
+This log was last generated on Wed, 24 Jan 2024 05:47:19 GMT and should not be manually modified.
+
+## 0.38.0
+Wed, 24 Jan 2024 05:47:19 GMT
+
+### Minor changes
+
+- Update dependencies.
+
+### Updates
+
+- Fix: Resolve selected arm common version when used on version enum members
+- Renamed template parameters for clarity and consistency.
+- Update dependencies
+
+## 0.37.1
+Wed, 20 Dec 2023 05:29:22 GMT
+
+### Patches
+
+- Add types to support lro scenarios
 
 ## 0.37.0
 Wed, 06 Dec 2023 19:47:28 GMT